+++ threedi_models_and_simulations/widgets/simulation_wizard.py
@@ -466,6 +466,7 @@ class BoundaryConditionsWidget(uicls_boundary_conditions, basecls_boundary_condi
             boundary_conditions_list = list(reader)
         error_msg = handle_csv_header(header)
         if error_msg is not None:
+            self.parent_page.parent_wizard.plugin_dock.communication.show_warn(error_msg)
             return None, None
         interpolate = (
             self.cb_interpolate_bc_1d.isChecked()