+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 
 - Fixes/enhancements: #616
 - Setting 1D initial substance concentration (#609)
+- Added user feedback for invalid boundary conditions file. (#624)
 
 
 3.14.1 (2024-09-25)